requirements.txt
asgiref==3.3.1
awesome-slugify==1.6.5
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
coverage==5.4
cryptography==3.3.1
defusedxml==0.6.0
Django==3.1.6
django-cors-headers==3.7.0
django-filter==2.4.0
django-js-asset==1.2.2
django-mptt==0.11.0
django-rest-auth==0.9.5
django-templated-mail==1.1.1
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
djoser==2.1.0
drf-yasg==1.20.0
idna==3.1
inflection==0.5.1
itypes==1.2.0
Jinja2==2.11.3
MarkupSafe==1.1.1
oauthlib==3.1.0
packaging==20.9
Pillow==8.1.0
psycopg2==2.8.6
pycparser==2.20
PyJWT==2.0.1
pyparsing==2.4.7
python-memcached==1.59
python3-openid==3.2.0
pytz==2021.1
regex==2020.11.13
requests-oauthlib==1.3.0
ruamel.yaml==0.16.12
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==4.0.3
sqlparse==0.4.1
Unidecode==1.1.2
uritemplate==3.0.1
urllib3==1.26.3
